pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dinov2-sim"
version = "0.1.0"
description = "Target image search in videos using DINOv2 features."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Generated by Assistant" }]
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "torchaudio>=2.2.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "scipy>=1.12.0",
    "imageio>=2.34.0",
    "imageio-ffmpeg>=0.4.9",
    "tqdm>=4.66.0",
    "hydra-core>=1.3.2",
    "typer>=0.9.0,<0.13.0",
    "click>=8.0.0,<8.2.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.1.0"]

[project.scripts]
dinov2-search = "dinov2_pipeline.cli:app"
